Add files via upload
This project uses the face_recognition library in Python to find a celebrity look-alike from a picture that a user uploads.

The face_recognition library, which is built using dlib’s state-of-the-art face recognition built with deep learning, is considered one of the simples libraries used for face recognition and manipulation. The model has an accuracy of 99.38% on the Labeled Faces in the Wild benchmark. With this library one can find faces, find and manipulate facial features, and identify faces in pictures. Users can also use this library with other Python libraries to do real-time face recognition.