Neural Style transfer using Deep Learning and Tensor flow 2
Neural Artistic Style Transfer finds a wide range of applications to fancily modify images. This field has so much influenced the technical world that many apps, such as Prisma, have received great craze amongst the users. In recent days, decent work has also been done in this area, which served as a holy grail to our project. The heart of this capability is the convolutional neural network, often used for modern image processing. The work of Leon A. Gatys, et al on ”A Neural Algorithm of Artistic Style”[1], has first come up with such an approach using a powerful pre-trained VGG19 model trained on the famous ImageNet database images.